+++ src/main/java/com/juridico/gestao/repositories/DadosProcessosRepository.java
@@ -3,7 +3,7 @@ package com.juridico.gestao.repositories;
 import com.juridico.gestao.Entity.Dadosprocesso;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DadosProcessosRepository extends JpaRepository<Dadosprocesso, Long> {
+public interface DadosProcessosRepository extends JpaRepository<Dadosprocesso, Integer> {
 
-    Dadosprocesso findByNumeroCnj(Long dadosProcesso);
+    Dadosprocesso findByNumeroCnj(Integer dadosProcesso);
 }
