+++ src/main/java/com/juridico/gestao/repositories/ProvisaoRepository.java
@@ -3,6 +3,6 @@ package com.juridico.gestao.repositories;
 import com.juridico.gestao.Entity.Provisao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProvisaoRepository extends JpaRepository<Provisao, Long> {
-    Provisao findByNumeroCnj(Long numeroCnj);
+public interface ProvisaoRepository extends JpaRepository<Provisao, Integer> {
+    Provisao findByNumeroCnj(Integer numeroCnj);
 }
