+++ src/main/java/com/juridico/gestao/repositories/RiscosRepository.java
@@ -4,6 +4,6 @@ import com.juridico.gestao.Entity.Provisao;
 import com.juridico.gestao.Entity.Riscos;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RiscosRepository extends JpaRepository<Riscos, Long> {
-    Riscos findByNumeroCnj(Long numeroCnj);
+public interface RiscosRepository extends JpaRepository<Riscos, Integer> {
+    Riscos findByNumeroCnj(Integer numeroCnj);
 }
